requirements.txt
python-dotenv
langchain-openai
langchain_community
chromadb
unstructured[all-docs]
streamlit
llama-index
llama-parse